+++ src/gui/qt/sidebar_plugin_name_delegate.h
@@ -30,7 +30,7 @@
 #include <QtWidgets/QStyledItemDelegate>
 
 namespace loot {
-qreal getSidebarRowHeight(bool inEditMode);
+int getSidebarRowHeight(bool inEditMode);
 
 class SidebarPluginNameDelegate : public QStyledItemDelegate {
   Q_OBJECT
